Add Customer View modules to bamazonCustomer.js; refactor; update README.md

+++ bamazon_schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE products (
   item_id INT(5) AUTO_INCREMENT NOT NULL,
   product_name VARCHAR(100) NOT NULL,
   department_name VARCHAR(100) NOT NULL,
-  price INT(10) NOT NULL,
+  price DECIMAL(8,2) NOT NULL,
   stock_quantity INT(5) NULL,
   PRIMARY KEY (item_id)
 );
